backend/requirements.txt
# Core FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# LangChain ecosystem
langchain>=0.1.0
langchain-openai>=0.0.2
langsmith>=0.0.77
openai>=1.3.0

# Database
supabase>=2.0.0

# Document processing 
PyPDF2>=3.0.0
pytesseract>=0.3.10
reportlab>=4.0.0

# HTTP and file handling
aiofiles>=23.2.0
requests>=2.31.0
httpx>=0.24.0,<0.25.0

# Web scraping for forms
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Vector DB
pinecone-client 